gitserver: Move vcs/git.ShortLog into the gitserver client
This is an example PR to show the approach we're thinking about adopting
in our gitserver client.
