+++ internal/gitserver/gitdomain/common.go
@@ -2,6 +2,7 @@ package gitdomain
 
 import (
 	"encoding/hex"
+	"fmt"
 	"strings"
 	"time"
 
@@ -102,4 +103,15 @@ type RefDescription struct {
 	Type            RefType
 	IsDefaultBranch bool
 	CreatedDate     time.Time
+}
+
+// A PersonCount is a contributor to a repository.
+type PersonCount struct {
+	Name  string
+	Email string
+	Count int32
+}
+
+func (p *PersonCount) String() string {
+	return fmt.Sprintf("%d %s <%s>", p.Count, p.Name, p.Email)
 }
